consolidate send and verify
