+++ src/packages/auth-ui-components/src/components/magic-link/index.ts
@@ -1,3 +1,2 @@
-export * from './verify';
-export * from './send';
+export * from './login';
 export * from './challenge';
