+++ src/packages/auth-ui-components/src/components/magic-link/login/index.ts
