--- src/packages/auth-ui-components/src/components/magic-link/send/graphql.ts
@@ -1,7 +1,0 @@
-import { DocumentNode, gql } from '@apollo/client';
-
-export const SEND_MAGIC_LINK_MUTATION: DocumentNode = gql`
-	mutation sendMagicLink($username: String!) {
-		sendMagicLink(username: $username)
-	}
-`;
