--- src/packages/auth-ui-components/src/components/magic-link/verify/graphql.ts
@@ -1,9 +1,0 @@
-import { DocumentNode, gql } from '@apollo/client';
-
-export const VERIFY_MAGIC_LINK_MUTATION: DocumentNode = gql`
-	mutation loginMagicLink($username: String!, $token: String!) {
-		result: loginMagicLink(username: $username, token: $token) {
-			authToken
-		}
-	}
-`;
