--- src/packages/auth-ui-components/src/components/magic-link/verify/index.ts
@@ -1,1 +1,0 @@
-export * from './component';
