--- src/packages/auth-ui-components/src/components/magic-link/verify/styles.module.css
@@ -1,38 +1,0 @@
-.wrapper {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	gap: 16px;
-	width: 250px;
-}
-
-.logo {
-	box-sizing: border-box;
-	padding: 40px;
-	width: 150px;
-}
-
-.titleContainer {
-	display: flex;
-	width: 100%;
-	justify-content: start;
-	font-size: 20px;
-	font-weight: 400;
-}
-
-.buttonContainer {
-	display: flex;
-	width: 100%;
-	justify-content: end;
-	padding: 40px 0;
-}
-
-.textInputField {
-	box-sizing: border-box;
-	height: 40px;
-	background: rgba(237, 232, 242, 0.02);
-	border: 1px solid rgba(237, 232, 242, 0.1);
-	border-radius: 6px;
-	padding: 10px;
-	width: 100%;
-}
