+++ src/packages/auth-ui-components/src/types.ts
@@ -1,0 +1,4 @@
+export enum AuthenticationMethod {
+	PASSWORD = 'pwd',
+	MAGIC_LINK = 'mgl',
+}
